:broom: Refactoring - banished new functions to container_uncertainty.py
